Rename TileData.h to Tile.h
