+++ OP2Utility.vcxproj.filters
@@ -69,9 +69,6 @@
     <ClInclude Include="src\Map\MapHeader.h">
       <Filter>Map</Filter>
     </ClInclude>
-    <ClInclude Include="src\Map\TileData.h">
-      <Filter>Map</Filter>
-    </ClInclude>
     <ClInclude Include="src\Map\TerrainType.h">
       <Filter>Map</Filter>
     </ClInclude>
@@ -169,6 +166,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="src\Map\Map.h">
+      <Filter>Map</Filter>
+    </ClInclude>
+    <ClInclude Include="src\Map\Tile.h">
       <Filter>Map</Filter>
     </ClInclude>
   </ItemGroup>
